moved import_module_from_file to inp_spec
