+++ input_parameters_format_specification.py
@@ -1,3 +1,8 @@
+from __future__ import print_function
+import sys
+import os
+import time
+import shutil
 import types
 from default_input_parameters import parameters_dict
 
@@ -40,19 +45,50 @@ def update_config_dict(config_dict, module, module_name, verbose=True):
 
     for parameter, default_value in optional_parameters.iteritems():
         if hasattr(module, parameter):
-            value = getattr(module, parameter)
             if parameter in config_dict and parameter != 't_ion':
                 raise PyECLOUD_ConfigException('Parameter %s is specified multiple times!' % parameter)
-            config_dict[parameter] = value
+            value = getattr(module, parameter)
         else:
+            if parameter in config_dict:
+                if parameter == 't_ion':
+                    continue
+                else:
+                    raise PyECLOUD_ConfigException('Parameter %s is specified multiple times!' % parameter)
             value = default_value
-            if parameter not in config_dict:
-                config_dict[parameter] = value
-            elif parameter == 't_ion':
-                continue
-            else:
-                raise PyECLOUD_ConfigException('Parameter %s is specified multiple times!' % parameter)
-
+        config_dict[parameter] = value
         if verbose:
             print('%s = %s' % (parameter, value))
+
+def import_module_from_file(module_name, file_name):
+    # Load any file as a python module. This function works for python2 and python3.
+    # See https://stackoverflow.com/questions/67631/how-to-import-a-module-given-the-full-path
+    # "One might think that python imports are getting more and more complicated with each new version."
+
+    # The importlib.util method requires the file name to end with '.py'.
+    # Also, loading the module from a temporary directory does not leave any .pyc files.
+    dir_name = '/tmp/PyECLOUD_%i' % os.getpid()
+    real_module_name = module_name + '_%f' % time.time()
+    if not os.path.isdir(dir_name):
+        os.mkdir(dir_name)
+
+    try:
+        new_file_name = dir_name+'/temp_file.py'
+        shutil.copy(file_name, new_file_name)
+
+        if sys.version_info.major == 2:
+            import imp
+            module = imp.load_source(real_module_name, new_file_name)
+
+        elif sys.version_info.major == 3:
+            import importlib.util
+            # In python3, imp.load_source works as well, however it does not return the exact same as the import statement.
+            # There are some superfluous contents in the namespace which collides with the inp_spec.assert_module_has_parameters method.
+
+            spec = importlib.util.spec_from_file_location(real_module_name, new_file_name)
+            module = importlib.util.module_from_spec(spec)
+            spec.loader.exec_module(module)
+    finally:
+        shutil.rmtree(dir_name)
+
+    return module
